add route system and fix folder structures
