+++ client/src/App.jsx
@@ -1,31 +1,62 @@
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
-import Login from './components/Login';
-import Home from './components/Home';
-import ShowProjects from './components/projects/ShowProjects';
-import CreateProject from './components/projects/CreateProject';
-import EditProject from './components/projects/EditProject';
-import ShowProjectsA from './components/assignament/ShowProjectsA';
-import AssignmentProject from './components/assignament/AssignmentProject';
+import Login from './components/pages/login/Login';
+import Home from './components/pages/home/Home';
+import ShowProjects from './components/pages/admin/ShowProjects';
+import CreateProject from './components/pages/admin/CreateProject';
+import EditProject from './components/pages/admin/EditProject';
+import ShowProjectsA from './components/pages/admin/ShowProjectsA';
+import AssignmentProject from './components/pages/admin/AssignmentProject';
+import AuthContextProvider from './context/AuthContext';
+import ProtectedRoute from './components/routes/ProtectedRoute';
+import LoginRoute from './components/routes/LoginRoute';
 
 function App() {
-  return (
-    <>
-      <BrowserRouter>
-        <Routes>
-          <Route path="/" element={<Login />}></Route>
-          <Route path="/panel" element={<Home />}></Route>
-          <Route path="/panel/projects" element={<ShowProjects />}></Route>
-          <Route
-            path="/panel/projects/create"
-            element={<CreateProject />}
-          ></Route>
-          <Route path="/panel/projects/edit/:id" element={<EditProject />} />
-          <Route path="/panel/assignment" element={<ShowProjectsA />}></Route>
-          <Route path="/panel/assignment/:id" element={<AssignmentProject/>}></Route>
-        </Routes>
-      </BrowserRouter>
-    </>
-  );
+	return (
+		<>
+			<BrowserRouter>
+				<AuthContextProvider>
+					<Routes>
+						<Route
+							path="/"
+							element={
+								<LoginRoute>
+									<Login />
+								</LoginRoute>
+							}
+						/>
+
+						<Route
+							path="/admin"
+							element={<ProtectedRoute requiredPermission="admin" />}
+						>
+							<Route index element={<Home />} />
+							<Route path="projects" element={<ShowProjects />} />
+							<Route path="projects/create" element={<CreateProject />} />
+							<Route path="projects/edit/:id" element={<EditProject />} />
+							<Route path="assignment" element={<ShowProjectsA />} />
+							<Route path="assignment/:id" element={<AssignmentProject />} />
+						</Route>
+
+						<Route
+							path="/pm"
+							element={<ProtectedRoute requiredPermission="pm" />}
+						>
+							<Route index element={<h1>Bienvenido Pm</h1>} />
+						</Route>
+
+						<Route
+							path="*"
+							element={
+								<LoginRoute>
+									<Login />
+								</LoginRoute>
+							}
+						/>
+					</Routes>
+				</AuthContextProvider>
+			</BrowserRouter>
+		</>
+	);
 }
 
 export default App;
