+++ client/src/components/common/Sidebar.jsx
@@ -1,12 +1,25 @@
-import '../assets/css/Sidebar.css'
-import { Link } from 'react-router-dom';
-import LogoArbusta from '../assets/icons/arbusta-icon.png';
-import DashboardIcon from '../assets/icons/dashboard-icon.svg'
-import BagIcon from '../assets/icons/bag-icon.svg'
-import UserIcon from '../assets/icons/user-icon.svg'
-import UsersIcon from '../assets/icons/users-icon.svg'
+import '../../assets/css/Sidebar.css'
+import { Link, useNavigate } from 'react-router-dom';
+import LogoArbusta from '../../assets/icons/arbusta-icon.png';
+import DashboardIcon from '../../assets/icons/dashboard-icon.svg'
+import BagIcon from '../../assets/icons/bag-icon.svg'
+import UserIcon from '../../assets/icons/user-icon.svg'
+import UsersIcon from '../../assets/icons/users-icon.svg'
+import { AuthContext } from '../../context/AuthContext';
+import { useContext } from 'react';
+
 
 export default function Sidebar() {
+
+  const { setUserData } = useContext(AuthContext);
+
+	const navigate = useNavigate();
+
+	const salir = () => {
+		setUserData({});
+		navigate('/');
+	};
+
   return (
     <>
       <aside className='sidebar-left'>
@@ -15,18 +28,18 @@ export default function Sidebar() {
           <p>ARBUSTA</p>
         </div>
         <nav className='nav'>
-          <Link to="/panel">
+          <Link to="/admin">
             <img src={DashboardIcon} alt="" />
             Dashboard
           </Link>
         </nav>
         <p>Proyectos</p>
         <nav className='nav'>
-          <Link to="/panel/assignment">
+          <Link to="/admin/assignment">
             <img src={ UsersIcon} alt="" />
             Asignar Empleados
           </Link>
-          <Link to="/panel/projects">
+          <Link to="/admin/projects">
             <img src={BagIcon} alt="" />
             Administrar Proyectos
           </Link>
@@ -38,9 +51,9 @@ export default function Sidebar() {
             Administrar usuarios
           </a>
         </nav>
-        <Link to="/">
+        <button onClick={salir}>
           Salir
-        </Link>
+        </button>
       </aside>
     </>
   )
